Moved to dijkstra_gif file.
