+++ tests/test_vis.py
@@ -19,19 +19,3 @@ def test_fragment_roll(monkeypatch):
 		data=path,
 		title="Testing fragment roll."
 	)
-
-# def test_dijkstra_gif():
-# 	# fp = os.path.dirname(here) + "/tests/static/Shuffled_Transcription_1.xml"
-# 	fp = "/Users/lukepoeppel/Messiaen/Encodings/Sept_Haikai/1_Introduction.xml"
-# 	import uuid
-# 	vis.dijkstra_gif(
-# 		filepath=fp,
-# 		part_num=0,
-# 		table=DecitalaHashTable(),
-# 		allow_subdivision=True,
-# 		title="Iterated Dijkstra on Sept Haïkaï (Bois)",
-# 		save_path=f"/Users/lukepoeppel/decitala/tests/dijkstra_gif_tests/dijkstra_{uuid.uuid4().hex}",
-# 		show=False
-# 	)
-	
-# print(test_dijkstra_gif())
